requirements.txt
#dependencias de lo que vamos a usar en el proyecto

fastapi
motor
pydantic
certifi
uvicorn
gunicorn
websockets
wrapt
uvloop
python-dotenv